% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info_data.R
\name{get_info_data}
\alias{get_info_data}
\title{flat_get_info_data Title}
\usage{
get_info_data(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
list
}
\description{
créer une fonction qui permettra d’avoir des informations sur un jeu de données passé en paramètre : dimensions et nom des colonnes dans une liste
}
\examples{
get_info_data(iris)
get_info_data(dplyr::starwars)
}
