% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mean_data.R
\name{get_mean_data}
\alias{get_mean_data}
\title{get_mean_data Title
Permet d'obtenir la moyenne de toutes les colonnes numeriques}
\usage{
get_mean_data(mydata)
}
\arguments{
\item{mydata}{dataframe}
}
\value{
1
}
\description{
get_mean_data Title
Permet d'obtenir la moyenne de toutes les colonnes numeriques
}
\examples{
get_mean_data(iris)

}
