% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydataset.R
\docType{data}
\name{mydataset}
\alias{mydataset}
\title{Starwars data, only Human species}
\format{
\subsection{\code{mydataset}}{

A tibble with 35 rows and 14 columns
}
}
\source{
\url{https://dplyr.tidyverse.org/reference/starwars.html}
}
\usage{
mydataset
}
\description{
A subset of data from the World Health Organization Global Tuberculosis
Report ...
}
\keyword{datasets}
