% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info_data.R
\name{get_info_data}
\alias{get_info_data}
\title{get_info_data Title}
\usage{
get_info_data(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
two messages
}
\description{
get_info_data Title
}
\examples{
get_info_data(iris)
}
