% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info_data.R
\name{get_info_data}
\alias{get_info_data}
\title{get_info_data}
\usage{
get_info_data(df)
}
\arguments{
\item{df}{un dataframe}
}
\value{
une liste contenant le nombre de lignes et de colonnes du dataframe
}
\description{
get_info_data
}
\examples{
get_info_data(iris)
get_info_data(mtcars)

}
