% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mean_data.R
\name{get_mean_data}
\alias{get_mean_data}
\title{get_mean_data}
\usage{
get_mean_data(df)
}
\arguments{
\item{df}{un dataframe}
}
\value{
une table
}
\description{
créer une fonction qui permettra d’obtenir une table avec la moyenne de toutes les variables numériques de votre jeu de donnée en paramètre
}
\examples{
get_mean_data(iris)
get_mean_data(mtcars)
}
